monitor/requirements.txt
requests>=2.28,<3
pyyaml>=6.0,<7
